+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add new `LocationIcon` React component.
+- Feature: Add new `LocationInput` React component.
+
 ## [8.0.0] - 2017-02-28
 
 Breaking change: